Add the `Exam` section in the Course Page. I updated the `Notes` section in the Course Page. I also added a Button component which makes it easy to create a toggle button group. I also added a ItemSubTitle.
Later on, I'll finish up the `Assignment` section in the Course page, then I'll do the `Figure` section. I'll work on the `Todo` section last because that looks like it will take the most time.

Lastly, I need to add a button to create a exam/note/assignments/figure. That will be simple, but I might save that for last. It'll just be a simple component that I'll tweak around here and there to work for everything.